refactor(api): extract diagnostic paper generation logic into generateStudentPaper
